Rename tests. Add tests for lea.
