+++ tests/test_machine_code_builder.py
